refactor: extract SwapsRecovery view and shared SeedWordInput component
